pyproject.toml
[project]
name = "meowth"
version = "0.1.0"
description = "Meowth Slack Bot - A Python Slack bot that responds to mentions with 'Meowth, that's right!'"
readme = "README.md"
authors = [
    { name = "Che Chia Chang", email = "chechia@maicoin.com" }
]
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0,<2.0.0",
    "llama-index>=0.9.0,<1.0.0",
    "llama-index-llms-azure-openai>=0.1.0",
    "tiktoken>=0.5.0",
    "llama-index-core>=0.12.52.post1",
    "llama-index-agent-openai>=0.4.12",
    "pydantic[yaml]>=2.12.4",
    "watchdog>=6.0.0",
    "pyyaml>=6.0.3",
    "pydantic-settings>=2.12.0",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
meowth = "meowth.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "dist", "build", ".venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "contract: Contract tests",
]
